requirements.txt
asgiref==3.2.10
Babel==2.8.0
Django==3.0.7
django-cors-headers==3.4.0
django-phonenumber-field==4.0.0
djangorestframework==3.11.0
phonenumbers==8.12.6
psycopg2-binary==2.8.5
pytz==2020.1
sqlparse==0.3.1
